# employee-management/docs/swagger.yaml
basePath: /
definitions:
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  controllers.ResponseMessage:
    properties:
      message:
        type: string
    type: object
...
  models.Department:
    properties:
      description:
        type: string
      employees:
        description: Liên kết với bảng Employees
        items:
          $ref: '#/definitions/models.Employee'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  models.Employee:
    properties:
      address:
        type: string
      cmnd:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      department_id:
        type: integer
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      position_id:
        type: integer
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing employees, departments, positions, and salaries.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Employee Management API
  version: "1.0"
paths:
  /api/v1/departments:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all departments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Department'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get all departments
      tags:
      - Department
    post:
      consumes:
      - application/json
      description: Add a new department record to the database
      parameters:
      - description: Department data
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/models.Department'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Create a new department
      tags:
      - Department
  /api/v1/departments/{department_id}/employees:
    get:
      consumes:
      - application/json
      description: Retrieve all employees in a specific department
      parameters:
      - description: Department ID
        in: path
        name: department_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Employee'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get employees by department
      tags:
      - Department
  /api/v1/departments/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a department record from the database
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Delete a department
      tags:
      - Department
    put:
      consumes:
      - application/json
      description: Update an existing department's information
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated department data
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/models.Department'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Update a department
      tags:
      - Department
...
swagger: "2.0"